chore(package.json): add some package.json fields
add a flag so that semantic-release doesn't try to publish to npm, to be removed when we can store

and view RSS feeds from the internet. Add a version number flag that is at 0.0.1 so that we have a

starting point to crafting our version numbers. Add babel-register to devDependencies which should

fix Travis